+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.3.0
 -----
 
+ * deprecated support for using integers as form names, use strings instead
  * deprecated using the `date_format`, `date_widget`, and `time_widget` options of the `DateTimeType` when the `widget`
    option is set to `single_text`
  * added `block_prefix` option to `BaseType`.